Restructure Cucumber test packages
